requirements.txt
Pillow
reportlab
pdf2image
PyPDF2